+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Both client libraries are pre-1.0, and they have separate versioning.
   - `Image.fromGcpArtifactRegistry` in typescript and `NewImageFromGcpArtifactRegistry` in golang.
 - Added `Secret.fromObject()` (JS) / `SecretFromMap()` (Go) to create a Secret from key-value pairs (like `from_dict()` in Python).
 - Added `name` field to `App`s, `Sandboxe`s, `Secret`s, `Volume`s, and `Queue`s.
+- Added support for `Function.getCurrentStats()` (JS) / `Function.GetCurrentStats()` (Go).
+- Added support for `Function.updateAutoscaler()` (JS) / `Function.UpdateAutoscaler()` (Go).
 
 ## modal-js/v0.3.16, modal-go/v0.0.16